Fix: 초기설정 버그 해결
