--- moamoa/GlobalStyle.jsx
@@ -1,11 +1,0 @@
-import { createGlobalStyle } from 'styled-components';
-import reset from 'styled-reset';
-
-const GlobalStyle = createGlobalStyle`
-${reset}
-:root{
-    
-}
-
-`;
-export default GlobalStyle;
