+++ moamoa/src/App.js
@@ -1,17 +1,15 @@
-import './App.css';
-import { createGlobalStyle } from 'styled-components';
-import reset from 'styled-reset';
-
-const GlobalStyle = createGlobalStyle`
-${reset}
-`;
+import { BrowserRouter } from 'react-router-dom';
+import AppRoutes from '../Router/AppRoutes';
+import GlobalStyle from './GlobalStyle';
 
 function App() {
   return (
-    <div className='App'>
-      <GlobalStyle />
+    <div>
+      <BrowserRouter>
+        <GlobalStyle />
+        <AppRoutes />
+      </BrowserRouter>
     </div>
   );
 }
-
 export default App;
