+++ moamoa/src/GlobalStyle.jsx
@@ -1,0 +1,11 @@
+import { createGlobalStyle } from 'styled-components';
+import reset from 'styled-reset';
+
+const GlobalStyle = createGlobalStyle`
+${reset}
+:root{
+    
+}
+
+`;
+export default GlobalStyle;
