+++ moamoa/src/Pages/Landing.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Landings from '../Components/Landings';
+
+export default function Landing() {
+  return (
+    <div>
+      <Landings />
+    </div>
+  );
+}
