+++ moamoa/src/Pages/Profile.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+
+export default function Profile() {
+  return (
+    <div>profile</div>
+  )
+
+}
